Allow building and pushing images on another branch
Before, all images are pushed with the SHA and the `latest` tag.

With this commit, pushing an image will tag it with the branch name as
well for easier use while developing.

Unfortunately, k8s doesn't support templating yet
(https://github.com/kubernetes/kubernetes/issues/23896) so you'll still
have to manually change the image tag in the `manifest.yaml` file back
and forth but it's definitely easier than before.

Now, the workflow is:

1. Work on image
2. Push with `make release-all`
3. Edit `manifest.yaml` to use images for your branch to test
4. Edit it back when making a PR.