Default name to namespace
Makes helm responsible for the namespace creation and deletion.
Previously, if you changed the name inside your yaml file it
will actually delete all your deployments and services and most
importantly - your PVCs, since they all belong in the namespace &
helm will just delete and recreate the namespace with new name.

This is almost never what you want, and so the name: value was just
asking for trouble. Given that there's also the helm release name,
this is doubly confusing. So let's just give it up :)

I'd have liked to just use Release.Name and make the Namespace
default to that too - but it seemed to take away a degree of
freedom in helm, since I can't just say 'default the namespace to
the name, but allow overrides'. I could theoretically do that by
creating the namespace ourselves, but that runs into all the
problems we previously mentioned :)