only set cookie secret env if specified
default behavior when cookie secret is null is already fine
because it will create jupyterhub_cookie_secret file in the db volume directory

cookieSecret is only necessary if there isn’t a volume mounted for the hub (i.e. sqlite db)

Even if not set, the only consequence is that rebooting the hub logs out users.

The non-sqlite case is not validated, however.
It would probably be nice to fail with an informative error if the db is external
and cookieSecret is not set.