Don't restart hub pod to pick up config change
Instead, whenever configmap changes, let's just restart the
hub *process*. We launch it with the watchdog manager, which
can restart processes when files change.

We try to have a 5s timeout to make sure that we don't restart
too many times for the same file. We also ignore files starting
with a '.' and directory changes for the same reason.