chart/heritage removed from template labels
A Deployment controls pods. These pods should not be considered to have a `heritage: tiller`, only the Deployment itself. If the template spec of these pods change, the Deployment will restart them, even though it is just a label to be added... Because of this, the "chart" label containing the chart version is removed, so a helm upgrade does not force a restart of the controlled pods.