Support offloading SSL to the proxy service
Rather than always having different target ports for the HTTP and
HTTPS frontends, setting proxy.https.type to offload will tell
the service to route ports 443 and 80 both to port 80 on the
backend.

For this to work, annotations need to be added to
proxy.service.annotations to attach certificates to the load
balancer.

Fixes #675