Instantly redirect to the progress page
@minrk pointed out that the slow redirection to the progress page is
related to the hub is waiting for the spawning to complete before it
redirects, as that may happen very quickly sometimes. But, on
kubernetes and clouds, we won't be able to insta-spawn a pod as we can
insta-spawn a process, so always redirecting to the progress page right
away makes sense.