user-scheduler added
The user-scheduler is great for use with a cluster autoscaler and bad 
without it. With the user-scheduler optionally enabled users will be 
packed to fill up one node at the time. This in turn tends to create 
totally unused nodes that are easy for a cluster autoscaler to scale 
down the unused node. Without a cluster autoscaler, it is just a waste 
of resources to get yourself unused nodes though, so only enable the 
user-scheduler if you have a cluster autoscaler.