Fix regarding image-puller & profileList - #1097
We forgot to check if the parent value was defined at one point. The go
templating that Helm relies on does not allow for a nice way to do this
with `and` functions as I understand it, see:
https://github.com/helm/helm/issues/3708

So, I opted for this functional solution even though it isnt very
elegant.

Closes #1097