deploy.sh renamed to publish-chart.sh
I think this is more descriptive. Previously I remember thinking that 
"deploy" referred to automatically keeping some z2jh chart updated, but 
in reality it was about publishing the chart to the 
jupyterhub/helm-chart repo.

This should hopefully reduce the time it takes to grasp what's going on 
for our developers.