Relax HSTS requirement on subdomains
Currently, if you have a JupyterHub at datahub.berkeley.edu with
HTTPS, you can *not* have a grafana.datahub.berkeley.edu without
HTTPS. This is because the default HSTS is set to 'includeSubdomains',
which means any subdomain must also enforce HTTPS.

This is a bit of an overreach, since the hub doesn't control any
other subdomain. This sets the default to be a bit more relaxed,
and allows people to turn it on if needed.