Clean up a line in the CI logs
This value was previously quoted with ", but when chartpress parses this
YAML and writes it back, the following change arise:

-     hstsIncludeSubdomains: "false"
+     hstsIncludeSubdomains: 'false'

So, by either making this value a boolean or a string with single quote
we get rid of this log entry coming from a `git diff`. I went with
single quotes for a minimalistic change.