Avoid +x complexity for cull_idle_servers.py
The cull_idle_servers.py script was started as an executable,
but the mounted file does not have the executable permissions
by default (0644), so i got a permission error.

Weirdly enough it can be resolved by adding mode: 0555 in the
volumeMount entry but that isn't part of the k8s API spec for
volumeMount! This was spotted by kubeval script.

I could set the defaultMode on the configmap volume definition
for the pod spec though. It would work, but it feels too ugly
as the configmap has more entries that shouldn't be considered
executable.

I decided to invoke it explicitly with python and mount it to
/etc/jupyterhub/cull_idle_servers.py instead. I figure that a
file mounted to /usr/local/bin should be executable, this isn't
and the folder named /srv/jupyterhub is already a mount point
for sqlite-pvc so I cannot use that and have a nested mount.