Support CD with git tags
TravisCI would make two builds if a commit and a tag referencing this
commit was pushed. This could lead to duplication of chart build and
publishing jobs. So, I disabled the tag triggered builds before to avoid
this, in case we ended up with duplicated entries or failure due to
existing entries etc.

In this PR, with chartpress 0.4.2 and the --long flag, we can let both
bulds finish. There are some key benefits.

1. We can now push a tag retrospectively now to trigger a build.
2. We avoid a failure to build a tag because it was pushed not as the
HEAD commit but some steps down the history.
3. We can still build on all pushed commits to master for dev releases.