[WIP] Replace kube-lego + nginx ingress with certbot + nginx
When we
[picked](https://github.com/jupyterhub/zero-to-jupyterhub-k8s/pull/229)
kube-lego to do automatic HTTPS in 2017, it seemed the easiest and
most kubernetes native way to do let's encrypt based HTTPS. It was
fairly simple, and worked pretty well for our use case. Users of
this chart could set up HTTPS without having to know anything about
certificates, challenges or let's encrypt. Life was good.

Life moves on, and so does code. kube-lego was deprecated in
favor of [cert-manager](https://github.com/jetstack/cert-manager),
which was far more general. It wasn't tied to Let's Encrypt,
and could do a lot more things. It was far more kubernetes native,
and slicker. However, this came with a massive increase in
complexity, which is justifyable if you need many certificates in
many situations. However, for the most common case of just one
certificate per hub, this was super massive overkill.

This PR replaces the current autohttps with autocertbot, a tiny
wrapper around the recommended upstream let's encrypt software -
[certbot](https://certbot.eff.org/). certbot requires persistent
storage under /etc/letsencrypt, and we would prefer to not. This
was the primary reason I picked kube-lego over certbot back in
the day. However, a small wrapper script that sync's the contents
of /etc/letsencrypt to a kubernetes secret when needed is small,
easy to reason about and solves all our problems.

This PR should be fully backwards compatible. autocertbot should
be able to renew certificates that kube-lego was handling before,
and issue new certificates as well. Since it is just wrapping
upstream certbot, there are fewer chances of a deprecation hitting
us hard.

Fixes #1448