Rewrite to use traefik instead of nginx
nginx proved more complicated than hoped for, primarily due to
lack of hot reloading. The two blockers were:

1. nginx won't start without SSL certs existing, but we won't get
   them until after nginx starts to use the webroot challenge!
2. When certificates get renewed, nginx needs a reload.

We could have fixed these, but I realized the reason we were not
using traefik for this was that it needs persistent disk to put its
let's encrypt config in. Since that is no longer a problem due to
the secret sync, I switched us over to traefik instead. *Much*
cleaner, simpler and straightforward!!!