DX: refactor helm template
Using `with` has become common enough that all users of helm templates 
needs to know how it works. It will not render anything if the value is 
falsy, but if it is truthy it will make the value accessible as `.` 
instead of needing to rewrite the same long reference again.