HSTS: Allow configuration of HSTS preload also
A preload value in the STS header indicates that the webserver wants the
browser to add this website to a list which should be considered to be
HTTPS only that can influence other browsers that haven't even visisted
this webserver before.

One can also manually add oneself to such list from
https://hstspreload.org/ if one complies with their requirements, of
which one is that a preload value in the STS header is specified.