DX: dev script / kind removed
The `dev` script was developed to make life easier, but it didn't help.
It was not robust enough and introduced shell script complexity, which
is the worst kind of complexity to introduce.

I'm reworking the local development setup, the removal of this script
and the coupling with the kind tool goes hand in hand with that.