Tweaking readiness/liveness probe: faster startup
With a readinessProbe that polls every ten seconds, we potentially need
to wait ten seconds after the pod is actually ready. By reducing this
number significantly, we avoid eigght potential seconds of waiting for
nothing.

At the same time, this could clutter logs. But, the proxy will not log
these requests, and I made a PR to JupyterHub to only log requests to
the health endpoint on the debug level.