Combine values.yaml data in hub-config to hub-secret
JupyterHub itself, authenticators, etc, all may want to use the values
we also use to render the Helm charts templates. So, we typically take
them av available in {{ .Values }}, split it apart, and render it to a
k8s configmap as well as a k8s secret.

This separation had little to no benefit and was hard to maintain. In
this commit I gather the {{ .Values }} into the hub-secret file. I also
configure c.JupyterHub.cookie_secret instead of passing that
information by mounting a k8s secrets key's value into a
JPY_COOKIE_SECRET environment variable on the hub pod.