Allow overriding hub container securityContext
This makes it possible to override the `securityContext`
for the hub container spec, similar to how the hub container
`resources` are overridable via `values.yaml`. By defualt the
`securityContext` is the same as before.

One use case for this is to be able to specify `capabilities`
in order to run py-spy [1] on the hub container.

One downside to this approach is that for anyone overriding the
hub container `securityContext` may need to duplicate the
`runAsUser` and `allowPrivilegeEscalation` entries which means
they could fall out of alignment with the upstream chart. So an
alternative to this full replacement approach is to add an
optional `extraContainerSecurityContext` entry similar to how
`extraVolumeMounts` works. The downside to that is it could
eventually conflict with what is not extra in the parent chart.
Another alternative is just adding a very specific and optional
`containerCapabilities` value which if set *only* sets the
`securityContext.capabilities` value in the container spec.

For simplicity and flexibility this opts for the full replacement
approach assuming the hub container `securityContext` will not change
frequently.

[1] https://github.com/benfred/py-spy#how-do-i-run-py-spy-in-kubernetes

Closes #1677