Be more explicit when setting HTTPS ports
For HTTPS, we want to expose port 443 on our service, but
only if it is 'turned on'. Unfortunately, in v 0.9, the
default was set to 'enabled' but it didn't act as enabled
unless hostnames were actually set. This exposed port 443
in proxy-public regardless of wether we were actually
set up for HTTPS or not. This caused weird issues with
AWS, since it was health checking all exposed ports. Since
we didn't have automatic HTTPS setup (since that needs
hostnames), this broke all installs on AWS!

This tries to be backwards compatible change, exposing
port 443 only if we *really* have https running. This
should fix issues with AWS, and be more 'correct'.

Manual HTTPS doesn't actually need hosts lists to be set,
so we remove that from our documentation.

Fixes #1637