hook-image-puller: -strict-check-duration flag added
It can be a problem to require a pod to be successfully scheduled on all
nodes. If we require this then chart upgrades when any node has
insufficient pods will fail, not because --wait part of the Helm upgrade
requires daemonset pods to be scheduled and ready because it doesn't,
but because the image-awaiter pod requires the desired number of pods to
be scheduled are scheduled and running.

This PR introduces a flag and sets a default value of 10 for it. Like
this, the criteria to successfully have awaited image pulling is to in
the initial ten seconds either have as many ready image puller pods as
is desired. The difference is that this will change after this initial
duration to something new, namely to wait only for the current number of
scheduled pods to become ready.