Set configurable minimal resource requests for image-puller daemonset pods
If a k8s cluster has a LimitRange resource, it may end up for example
allocating 100m CPU for containers which is a waste of resources. This
would be a waste as the running container only does `sh -c "Done!"`
anyhow. While there is image pulling happening, that is done before the
actual container starts so its worth noting we don't need any resources
specified for that.

Another reason for specifying resources, and making them configurable,
is that sometimes there is a ResourceQuota resource that requires the
resources to be explicitly specified on pods.

Related:
- Default CPU requests - https://kubernetes.io/docs/tasks/administer-cluster/manage-resources/cpu-default-namespace/
- LimitRange https://kubernetes.io/docs/concepts/policy/limit-range/
- ResourceQuota - https://kubernetes.io/docs/concepts/policy/resource-quotas/