Relax hook-image-puller to make upgrades more reliable
This configuration makes the hook-image-awaiter start comparing the
number of scheduled image-puller pods with the number of ready
image-puller pods, instead of comparing the number of desired
image-puller pods with the ready number of image-puller pods, after a
duration of 10 seconds.

This helps anyone using the hook-image-puller to avoid failures
of running `helm upgrade` that can happen if a node is full packed with
pods and no more pods can be scheduled on it, because then the
image-awaiter would get stuck waiting forever on an image pulling pod
that never successfully schedules to do the actual pulling.