Match initialDelaySeconds for all hub probes
Right now, k8s starts checking if the hub is up as soon
as it starts. If the hub isn't fully up in about 10s,
it *restarts the hub*, which then has top start initializing
all over again! If you're lucky, the hub will actually be
able to finish initializing before CrashLoopBackoff, but
very easy to be stuck in CrashLoopBackoff with even a moderately
large hub.