CI: fix RTD builds on push to master
It appears that when we build the docs in the PR preview, it works. But,
when we push to master, it doesn't any more. It appears that using
pydata_sphinx_theme to build PDFs and/or ePub files makes it crash if we
using sphinx3 rather than sphinx2. Since a doc dependency that
(executablebooks/myst-parser) previously was limited to sphinx 2 but not
longer is, the PDF / ePub build made our builds when pushing to master
fail.

I'd like to avoid maintaining these things, and would like to stop
building PDFs and ePub files. I don't know if they are used, but I'd
like to focus attention to https://z2jh.jupyter.org anyhow I think.