Added config prePuller.pullProfileListImages
With a large list of images in the singleuser.profileList, users ended
up with delays caused by images being pulled for so long time that it
blocked what was actually important to pull to get started properly.

This option allows for a long singleuser.profileList with many different
images to be setup, while only pulling the default image, or optionally
those configured under prePuller.extraImages.