Revert recent removal of image-pulling related to cloudMetadata blocker
I previously removed this logic because it was messy to implement the
deprecation logic, and the container would be so very small anyhow. But,
I realized that the size may only be a smaller part if this matters or
not. Each image needs to be pulled, and if they end up being pulled in
the wrong order, it can be very troublesome.

What could happen without this change, is that the image puller pods
could spend time pulling multiple images, then a user pod arrives to the
node and its image is already pulled and it only needs to pull the small
cloudMetadata blocker image, but, pulling that is blocked currenlty.
Because images are pulled in a sequence as I understand it, so the user
pod arriving may end up needing to wait for its little initContainer
image to be pulled and that is just bad.

So, the it matters if we pull this small image, and it matters when we
pull it. We should pull it the first thing we do.