Provide a way to disable default hub configmap
The jupyterhub chart is ultra-configurable and good for deploying
my other jupyterhub environments.  But it provides no way to not
use the default jupyterhub_config.py via the configmap.  I want
to be able to use the extraConfigMaps to mount my own configmap
that contains the jupyterhub_config.py and not include the original
jupyterhub_config.py.