netpol: allow internamespace communication
Our network policies only allowed pods within the local namespace to
communicate with the hub/proxy/singleuser-server pods if they had a
certain label. But, we want any pod in any namespace with the relevant
access labels to be allowed to communicate with the associated pod, not
only pods within the local namespace.

This commit makes that possible!