Allow setting extra public ports on proxy-public
Now that we have proxy.traefik.extra(Static|Dynamic)Config,
we can configure traefik to do more proxying. Specifically,
I want it to proxy ssh traffic to an instance of
[jupyterhub-ssh](https://github.com/yuvipanda/jupyterhub-ssh).

This works fine, but we need to have proxy-public expose
port 22 to the world. Currently this is impossible -
this commit allows adding additional ports here.