Allow exposing extra ports in traefik deployment
Follow-up to #1852. Useful to explicitly expose ports
in deployment too - otherwise you can't use port names like
`ssh` in service definitions. To allow ingress on a port named
`ssh` in traefik right now, you have to know it is listening
on port 8022, and set the following:

```yaml
proxy:
  autohttps:
    networkPolicy:
      allowedIngress: [http, https, 8022]
```

This requires knowledge that `8022` is the ssh port. Instead,
it would be nice to specify `ssh` instead - and that requires
defining that in the deployment. This is true for other places
too - like the service definition.