ci: use socket.connect instead of wget to test connectivity
I got 403 connection errors using wget and passing the host header
explicitly from the jupyter.org webserver using the GitHub runners but
not if running it locally. Instead of debugging that difference, I'll
transition to use a more fundamental test which is what matters. Can we
open a TCP connection to the remote server at all? To do that, I'm now
using Python itself.