auth rework: manage auth.state.cryptoKey / JUPYTERHUB_CRYPT_KEY
I updated the logic for checking if we had a valid configuration as
well using the Helm template helpers I introduced as I concluded it was
needed to remain sane while checking nested config that may be null at
any place.

I deleted the entry in secret.yaml as it is only to be used to set an
env variable for use by JupyterHub in the hub pod, which in turn is fine
to delete because its only used by JupyterHub's CryptKeeper class which
is responsible for managing encrypt/decrypt operations for JupyterHub.