ci: stop using kubeval and rely on helm template --validate
kubeval and the json schemas is a project with too little maintainers,
so now we can't validate against k8s 1.19 or 1.20. Thankfully, we have
helm template --validate now that allow us to use a k8s api-server to
validate that our rendered templates are valid k8s resources.

By running a helm template --validate step with the
lint-and-validate-values.yaml we can ensure that they are valid.