schema: allow null for some but not all values
I decided that schema native values should always be set and not allowed
to be null. But, for configuration values that are pass through
configuration to JupyterHub and KubeSpawner for example, it made more
sense to let them be null as the default of the Helm chart may be to not
set it at all unless explicitly passed by a user.