Don't restart the proxy pod with each deploy
Restarting the proxy pod causes current user sessions to
drop until the pod restarts, the hub notices it, and sets up
all the routes again. On a hub with a few thousand users, this
can be minutes. Instead, we provide the proxy pod with just the
one key it needs, so it'll restart only if that changes.