Remove helmlint from pre-commit
It was causing trouble for us as it couldn't be automatically installed
with `helm` as a dependency from where it was defined. Due to this and
the fact that I believe `helm lint` captures very few errors and that we
have a CI check for this and more already, I figure it is better we just
remove it because otherwise we can't use the pre-commit.ci service.