ci: fetch stable/dev releases using helm show to avoid cache issues
Previously we fetched the latest stable and dev release from metadata we
have maintained, but doing that and then using `helm install` of that
version made us use two sources of data, where one could be relatively
outdated due to hitting a cache while the other may not.

Using `helm show`, we rely on the helm chart repository website
returning a index.yaml listing versions etc both when checking whats the
latest versions, and when installing - like that we avoid a cache issue
I think.